+++ S2ZonaConfigurator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog 
 
+## [0.0.3]
+- Add a feature to copy additional files (like assets) to the output directory
+- Handle more cases when the mod should be marked as failed
+
 ## [0.0.2]
 - Add compression to the generated .pak file
 - AES key is optional